requirements.txt
matplotlib==3.7.1
pandas==1.5.2
psycopg2_binary==2.9.5
selenium==4.8.2
webdriver_manager==3.8.5
